backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
Automat==25.4.16
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6  # có thể không cần thiết trên Linux, nếu có thể loại bỏ
comm==0.2.2
constantly==23.10.4
contourpy==1.3.1
crawler==0.0.2
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
dask==2025.3.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Django==5.1.3
djangorestframework==3.15.2
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2025.3.0
greenlet==3.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itemadapter==0.11.0
itemloaders==1.3.2
jedi==0.19.2
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
locket==1.0.0
lxml==5.3.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
multidict==6.4.3
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parsel==1.10.0
parso==0.8.4
partd==1.4.2
pillow==11.0.0
platformdirs==4.3.6
playwright==1.51.0
prompt_toolkit==3.0.50
propcache==0.3.1
Protego==0.4.0
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
pyee==12.1.1
Pygments==2.19.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
queuelib==1.8.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
scikit-learn==1.6.1
scipy==1.15.2
Scrapy==2.12.0
scrapy-playwright==0.0.43
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
SQLAlchemy==2.0.40
sqlparse==0.5.2
stack-data==0.6.3
threadpoolctl==3.6.0
tldextract==5.2.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tweepy==4.15.0
Twisted==24.11.0
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.2.3
w3lib==2.3.1
wcwidth==0.2.13
wheel==0.45.1
yarl==1.20.0
zope.interface==7.2
